Make Javacomplete works with python 2 or 3.
[Problem]
Currently javacomplete plugin relies on Python to do some of its work.
Unfortunately it strictly rely on Python 2 even though the code can run on
Python 3 without any problem.

[Solution]
Make the code first check which Python is available and the first one.

[Notes]
This change essentially remove the use of pyeval due the fact Vim does not
support this function if only Python 3 is available.

Also, I put a check on the beginning of the file to guarantee that we don't
proceed if no Pythno is available.

Signed-off-by: Thiago Alves <talk@thiagoalves.com.br>