Renamed Py* commands to JavacompletePy*.
[Problem]
Because we use a very generic name for our Python commands (Py and Pyfile) there
is a chance other plugins or the user might have already created those commands
and overriding them might cause unwanted effects for the user.

[Solution]
Use a more "unique" name for our commands to minimize the change of conflict.
Although these commands will be globally available, the user would have to
create a "JavacompletePy" command in order to have problem.